Replace 2022/08/18 privacy policy with 2023/01/01 privacy policy
