--- src/frontend/pages/privacy-preview.tsx
@@ -1,5 +1,0 @@
-import PrivacyPreview from "src/views/PrivacyPreview";
-
-const Page = (): JSX.Element => <PrivacyPreview />;
-
-export default Page;
