+++ src/frontend/src/views/Privacy/index.tsx
@@ -16,6 +16,7 @@ import {
 import { PageContent } from "../../common/styles/mixins/global";
 import {
   ContentRow,
+  H2AnchorLink,
   SectionRow,
   Table,
   TopRow,
@@ -27,7 +28,7 @@ const PrivacyPolicy = (): JSX.Element => {
     <>
       <Title>
         <H1>Chan Zuckerberg GEN EPI (formerly Aspen) Privacy Policy</H1>
-        <H4>Last Updated: August 18, 2022</H4>
+        <H4>Last Updated: January 1, 2023</H4>
       </Title>
       <P>
         The Chan Zuckerberg Initiative Foundation, a 501(c)(3) nonprofit private
@@ -40,7 +41,7 @@ const PrivacyPolicy = (): JSX.Element => {
         by CZ GEN EPI Users (collectively &quot;Users&quot; or &quot;you&quot;,
         ex: registered public health officials at state and/or county level
         Departments of Public Health (&quot;<B>DPH</B>&quot;), other public
-        health researchers), and how we use, share, and protect that
+        health researchers), and how we use, disclose, and protect that
         information.
       </P>
       <H3>About CZ GEN EPI</H3>
@@ -58,13 +59,14 @@ const PrivacyPolicy = (): JSX.Element => {
         Raw Sequence Data (as described below) as well as information about
         those sequences, such as the date the sample was collected (&quot;Sample
         Metadata&quot; as further defined below -- Raw Sequence Data and Sample
-        Metadata together make &quot;Upload Data&quot;). Any human genetic data
-        contained within the Raw Sequence Data is filtered out and deleted
-        following upload, leaving genomic data only about the pathogen. This
-        pathogen genomic data is then analyzed in order to identify the
-        normally-occurring genetic mutations that make up each pathogen sample’s
-        unique genetic &quot;barcode.&quot; This barcode can then be used to
-        identify strains, variants, and relationships between samples.{" "}
+        Metadata together make &quot;Upload Data&quot;). In the event any human
+        genetic data is contained within the Raw Sequence Data, it is filtered
+        out and deleted following upload, leaving genomic data only about the
+        pathogen. This pathogen genomic data is then analyzed in order to
+        identify the normally-occurring genetic mutations that make up each
+        pathogen sample’s unique genetic &quot;barcode.&quot; This barcode can
+        then be used to identify strains, variants, and relationships between
+        samples.{" "}
         <B>
           By default, these analytical outputs will be visible to the User that
           uploaded the Sample and other members of the User’s organization
@@ -113,9 +115,10 @@ const PrivacyPolicy = (): JSX.Element => {
           internal identifiers.
         </ListItem>
         <ListItem>
-          Human genetic data with uploaded data is processed only so we can
-          filter out and permanently delete it. We do not keep this non-pathogen
-          genomic data and it’s not necessary to operate the tool.
+          In the event any human genetic data is uploaded by you, that data is
+          processed only so we can filter out and permanently delete it. We do
+          not keep this non-pathogen genomic data and it’s not necessary to
+          operate the tool.
         </ListItem>
         <ListItem>
           <span>
@@ -166,8 +169,8 @@ const PrivacyPolicy = (): JSX.Element => {
             <td>
               <P>
                 Upon upload, Raw Sequence Data is processed through our data
-                pipeline and all human genetic information is filtered out and
-                deleted.
+                pipeline and all human genetic information (if any) is filtered
+                out and deleted.
               </P>
               <P>
                 We use the remaining data, with Sample Metadata, to create the
@@ -184,6 +187,10 @@ const PrivacyPolicy = (): JSX.Element => {
                 Other than as specifically requested by you, such as to debug an
                 issue, staff working on CZ GEN EPI never access this data.
               </P>
+              <P>
+                This data is temporarily disclosed to our service providers (ex:
+                AWS) in order to operate the service.
+              </P>
             </td>
             <td rowSpan={4}>
               <P>
@@ -193,8 +200,9 @@ const PrivacyPolicy = (): JSX.Element => {
                 <NewTabLink href="mailto:hello@czgenepi.org">
                   hello@czgenepi.org
                 </NewTabLink>{" "}
-                and we will fulfill the request within 60 days or sooner
-                depending on applicable privacy laws.
+                and we will fulfill the request within the timelines set forth
+                under applicable privacy laws. You can learn more about your
+                choices in the Section titled, “Choices About Your Data.”
               </P>
               <P>
                 Please be aware, however, that we cannot delete any Pathogen
@@ -285,7 +293,7 @@ const PrivacyPolicy = (): JSX.Element => {
             </td>
             <td>
               Analytical results you create are visible to other users in your
-              Group, as well as third party entities, that your Group is visible
+              Group, as well as third party entities that your Group is visible
               to.
             </td>
           </ContentRow>
@@ -302,8 +310,9 @@ const PrivacyPolicy = (): JSX.Element => {
             <td>
               Data about researchers with CZ GEN EPI accounts such as name,
               email, institution, basic information about how they are using CZ
-              GEN EPI, and information provided for user support (ex: resolving
-              support requests).
+              GEN EPI, and information provided through surveys, during an
+              in-person event, or for user support (ex: resolving support
+              requests).
             </td>
             <td>
               We use this data only to operate, secure, and improve the CZ GEN
@@ -334,8 +343,9 @@ const PrivacyPolicy = (): JSX.Element => {
               <NewTabLink href="mailto:hello@czgenepi.org">
                 hello@czgenepi.org
               </NewTabLink>{" "}
-              and we will fulfill the request within 60 days or sooner depending
-              on applicable privacy laws.
+              and we will fulfill the request within the timelines set forth
+              under applicable privacy laws. You can learn more about your
+              choices in the Section titled, “Choices About Your Data.”
             </td>
           </ContentRow>
           <ContentRow>
@@ -343,16 +353,31 @@ const PrivacyPolicy = (): JSX.Element => {
               <B>Device and Analytics Data</B>
             </td>
             <td>
-              Device Data (ex: browser type and operating system) and Analytics
-              Information (ex: links within CZ GEN EPI you click on and how
-              often you log into CZ GEN EPI) includes basic information about
-              how Users and Visitors are interacting with CZ GEN EPI.
+              Device Data (ex: browser type and operating system, and IP
+              address) and Analytics Information (ex: links within CZ GEN EPI
+              you click on and how often you log into CZ GEN EPI) includes basic
+              information about how Users and Visitors are interacting with CZ
+              GEN EPI and a general location (i.e. country) of where they are
+              accessing the Services.
             </td>
             <td>See above.</td>
             <td>See above.</td>
           </ContentRow>
         </tbody>
       </Table>
+    </>
+  );
+
+  const renderDataCZGenEpiCollects = () => (
+    <>
+      <H2AnchorLink id="privacy-data-collection">
+        Data CZ GEN EPI Collects
+      </H2AnchorLink>
+      <P>
+        We will never share your personal information for behavioral advertising
+        purposes nor will we ever sell your personal information. We do collect,
+        use, and retain your data as follows:
+      </P>
     </>
   );
 
@@ -390,14 +415,12 @@ const PrivacyPolicy = (): JSX.Element => {
           <span>
             <ListItemLabel>Raw Sequence Data:</ListItemLabel> &quot;Raw Sequence
             Data&quot; is genomic sequence data, including both host and
-            pathogenic data. As part of the process of processing this data
-            uploaded to CZ GEN EPI, any identifiable human genetic data is
-            filtered and removed. This means that CZ GEN EPI should <B>not</B>{" "}
-            contain any human sequence data. Note that if there are no issues
-            identified with the corresponding Pathogen Consensus Genome, the Raw
-            Sequencing Data will be permanently deleted from our backend after
-            90 days. We encourage Users to submit raw reads to the Sequencing
-            Read Archive for long-term storage.
+            pathogenic data. In the event any identifiable human genetic data is
+            uploaded, it will be filtered out and removed. Note that if there
+            are no issues identified with the corresponding Pathogen Consensus
+            Genome, the Raw Sequencing Data will be permanently deleted from our
+            backend after 90 days. We encourage Users to submit raw reads to the
+            Sequencing Read Archive for long-term storage.
           </span>
         </ListItem>
         <ListItem>
@@ -430,9 +453,20 @@ const PrivacyPolicy = (): JSX.Element => {
         and we will address it.
       </P>
       <P>
+        <UnderLineHeader>Sensitive Personal Information</UnderLineHeader>
+      </P>
+      <P>
+        In the event Raw Sequence Data from a human host is uploaded by a User
+        to CZ GEN EPI, it is sensitive personal information (genomic sequence
+        data), which will only be used to operate and provide the requested
+        service to you. As mentioned throughout this Privacy Policy, any human
+        genetic information contained in your upload is filtered out and deleted
+        and only used to create the Pathogen Consensus Genome.
+      </P>
+      <P>
         <UnderLineHeader>How We Use Upload Data</UnderLineHeader>
       </P>
-      <P>Upload Data is used for the following purposes:</P>
+      <P>Upload Data is used for the following business purposes:</P>
       <List>
         <ListItem>
           To provide Users and their Groups with a &quot;Pathogen Consensus
@@ -454,24 +488,28 @@ const PrivacyPolicy = (): JSX.Element => {
         </ListItem>
       </List>
       <P>
-        We do not own Upload Data and will never sell it. As mentioned above,
-        your Upload Data <B>will</B> be visible within your Group.
+        We do not own Upload Data and we will never sell your data nor will we
+        share your personal information for behavioral advertising purposes. As
+        mentioned above, your Upload Data <B>will</B> be visible within your
+        Group.
       </P>
       <P>
-        <UnderLineHeader>How We Share Upload Data</UnderLineHeader>
+        <UnderLineHeader>How We Disclose Upload Data</UnderLineHeader>
       </P>
       <P>
         Raw Sequence Data and Sample Metadata are shared back to the Users that
         uploaded the data, as well as other Users within the same organization
-        (your &quot;Group&quot;). Please note that while the Raw Sequence Data
-        is temporarily visible to other members of your Group, this data is not
-        retained on the CZ GEN EPI platform.
+        (your &quot;Group&quot;). This data is also disclosed to service
+        providers (ex: AWS) in order to store the data. Please note that while
+        the Raw Sequence Data is temporarily visible to other members of your
+        Group, this data is not retained on the CZ GEN EPI platform and is
+        permanently deleted from our backend after 90 days.
       </P>
       <P>
-        We may also share your Pathogen Consensus Genomes (whether uploaded by
-        you or generated by us) and/or analytical outputs with third parties in
-        accordance with the provisions of your organization’s policies and/or as
-        required by law. For example, certain Users in California currently
+        We may also disclose your Pathogen Consensus Genomes (whether uploaded
+        by you or generated by us) and/or analytical outputs with third parties
+        in accordance with the provisions of your organization’s policies and/or
+        as required by law. For example, certain Users in California currently
         allow the California Department of Public Health (&quot;CDPH&quot;) to
         access data from their Group. Where such access is allowed by Groups,
         the third party can access this data through their own CZ GEN EPI
@@ -487,7 +525,7 @@ const PrivacyPolicy = (): JSX.Element => {
       </P>
       <P>
         Pathogen Consensus Genomes, whether uploaded by you or generated by CZ
-        GEN EPI, will be shared by us with public repositories (as set out
+        GEN EPI, will be disclosed by us with public repositories (as set out
         below) unless you choose to mark this information as
         &quot;private.&quot; In the event that the Pathogen Consensus Genome is
         created by us, it will automatically be marked as private if the
@@ -495,13 +533,19 @@ const PrivacyPolicy = (): JSX.Element => {
       </P>
       <P>
         <UnderLineHeader>
-          What’s our legal basis to use and share Upload Data?
+          What’s our legal basis to use and disclose Upload Data?
         </UnderLineHeader>
       </P>
       <P>
-        Data uploaded to CZ GEN EPI by Users should <B>always</B> be anonymous.
-        The pathogen genome does not contain personal data, as it cannot be
-        personally linked with an identifiable individual.
+        To the extent that the GDPR applies, we rely on the following legal
+        bases to use and disclose personal data in Upload Data:
+      </P>
+      <P>
+        The Raw Sequence Data is processed to support research regarding the
+        study of infectious disease with the potential to benefit global public
+        health. Upon upload, human genetic information (if any) is filtered out
+        to create a Pathogen Consensus Genome which is used to advance
+        scientific research.
       </P>
       <P>
         In the rare event that human genetic data is not successfully deleted in
@@ -510,6 +554,10 @@ const PrivacyPolicy = (): JSX.Element => {
         legitimate interest, and in the legitimate interests of CZB and CZI LLC,
         in order for us to ensure that no personal data is contained within the
         genomic data stored on CZ GEN EPI.
+      </P>
+      <P>
+        The remaining pathogen genome does not contain personal data, as it
+        cannot be personally linked with an identifiable individual.
       </P>
     </>
   );
@@ -521,13 +569,13 @@ const PrivacyPolicy = (): JSX.Element => {
         created from Upload Data
       </H2>
       <P>
-        If you have uploaded Raw Sequence Data, we first strip any human reads,
-        and then generate a Pathogen Consensus Genome by mapping the remaining
-        sequencing reads to a pathogen-specific reference genome. These
-        Consensus Genomes are the foundational unit of analysis for genomic
-        epidemiology. If you have submitted Pathogen Consensus Genomes as Raw
-        Sequence Data, we simply align it to the appropriate pathogen reference
-        genome.
+        If you have uploaded Raw Sequence Data, we first strip any human reads
+        (if any), and then generate a Pathogen Consensus Genome by mapping the
+        remaining sequencing reads to a pathogen-specific reference genome.
+        These Consensus Genomes are the foundational unit of analysis for
+        genomic epidemiology. If you have submitted Pathogen Consensus Genomes
+        as Raw Sequence Data, we simply align it to the appropriate pathogen
+        reference genome.
       </P>
       <P>
         CZ GEN EPI also gives you the ability to create new analytical outputs
@@ -567,9 +615,10 @@ const PrivacyPolicy = (): JSX.Element => {
       </H2>
       <P>
         CZ GEN EPI also collects information about Users in order to offer the
-        Service. Other than basic information required to create an account
-        (e.g. email address, name, Group affiliation), the User determines what
-        information they want to upload onto CZ GEN EPI.
+        Service. CZ GEN EPI collects basic identifiers required to create an
+        account (e.g. email address, name, Group affiliation), Device and
+        Analytics Data, information from Cookies, and information Users decide
+        they want to upload onto CZ GEN EPI.
       </P>
 
       <UnderLineHeader>What We Collect</UnderLineHeader>
@@ -579,28 +628,30 @@ const PrivacyPolicy = (): JSX.Element => {
           <span>
             <ListItemLabel>User Data.</ListItemLabel>
             User Data is any information we collect from a User about that User
-            (&quot;User Data&quot;). It may include information necessary to
-            create or access your account such as your name, email, Group name
-            and contact email, and login credentials.
+            such as identifiers and professional or employment-related
+            information (&quot;User Data&quot;). It may include information
+            necessary to create or access your account such as your name, email,
+            Group name and contact email, and login credentials. This could also
+            include information we collect if you decide to respond to a
+            voluntary survey or join us for an in-person event.
           </span>
         </ListItem>
         <ListItem>
           <span>
-            <ListItemLabel>Analytics.</ListItemLabel>
+            <ListItemLabel>Device and Analytics Data.</ListItemLabel>
             When Users visit or use our Service, we may automatically collect
-            some information so that we can understand the way in which our tool
-            is being used. We may collect some Device Data or Analytics
-            Information in order to do this. &quot;Device Data&quot; includes
-            information about your browser type and operating system, IP address
-            and/or device ID. &quot;Analytics Information&quot; relates to any
-            of your requests, queries, or use of the Services, such as the
-            amount of time spent viewing particular web pages. We use Analytics
-            Information in accordance with our legitimate interests. Any data
-            which we collect for analytics purposes will be stored in a
-            de-identified and aggregated manner wherever possible; any analytics
-            data that is not able to be aggregated and de-identified will not be
-            shared beyond the CZ GEN EPI team and will be stored for no longer
-            than is necessary.
+            some information which includes internet or other electronic network
+            activity and geolocation data so that we can understand the way in
+            which our tool is being used and where our Users are accessing the
+            tool from (i.e., country). We may collect some Device Data or
+            Analytics Information in order to do this. &quot;Device Data&quot;
+            includes information about your browser type and operating system,
+            IP address and/or device ID. &quot;Analytics Information&quot;
+            relates to any of your requests, queries, or use of the Services,
+            such as the amount of time spent viewing particular web pages. We
+            use Analytics Information in accordance with our legitimate
+            interests. Any data which we collect for analytics purposes will be
+            stored in a de-identified and aggregated manner wherever possible.
           </span>
         </ListItem>
         <ListItem>
@@ -628,7 +679,7 @@ const PrivacyPolicy = (): JSX.Element => {
       </P>
       <P>
         User Data will be used to operate, secure, and improve the Services.
-        This means the following purposes:
+        This means the following business purposes:
       </P>
       <List>
         <ListItem>
@@ -641,8 +692,9 @@ const PrivacyPolicy = (): JSX.Element => {
         </ListItem>
         <ListItem>To respond to your inquiries and requests.</ListItem>
         <ListItem>
-          To analyze broadly how Users are using CZ GEN EPI so we can optimize
-          and improve it.
+          To analyze broadly how Users are using CZ GEN EPI and the general
+          location (i.e., country) of where Users are accessing the Service so
+          we can optimize and improve it.
         </ListItem>
         <ListItem>
           To protect the security and integrity of CZ GEN EPI.
@@ -654,6 +706,10 @@ const PrivacyPolicy = (): JSX.Element => {
         </UnderLineHeader>
       </P>
       <P>
+        To the extent that the GDPR applies, we rely on the following legal
+        bases to use and disclose personal data in User Data:
+      </P>
+      <P>
         We (along with CZB and CZI LLC) have a legitimate interest in using
         personal data within User Data in the ways described in this Privacy
         Policy to operate, secure, and improve CZ GEN EPI. This allows us to
@@ -661,93 +717,9 @@ const PrivacyPolicy = (): JSX.Element => {
         research regarding the study of infectious disease with the potential to
         benefit global public health.
       </P>
-    </>
-  );
-
-  const renderVendorAndThirdPartyPolicy = () => (
-    <>
-      <H2>
-        <Number>4.</Number>Vendors and Other Third Parties
-      </H2>
       <P>
-        CZB, CZIF, and CZI LLC collaborate closely in order to build, design,
-        and operate CZ GEN EPI so that it can be as useful as possible to
-        researchers and the public health community. CZB and CZIF provide
-        scientific and data analysis leadership and CZI LLC focuses on
-        maintaining CZ GEN EPI’s infrastructure, security, and compliance. The
-        three parties are all data controllers for CZ GEN EPI and will all only
-        use data as described in this Privacy Policy.
-      </P>
-      <P>
-        We also use service providers, such as a database provider and analytics
-        providers, to support and improve the operation of CZ GEN EPI. These
-        service providers are data processors and their use is limited to the
-        purposes disclosed in this Privacy Policy. For more information about
-        our use of service providers, please see our{" "}
-        <NewTabLink href={ROUTES.PRIVACY_DATA_SHARING_FAQ}>
-          Privacy &amp; Data Sharing FAQ
-        </NewTabLink>
-        .
-      </P>
-      <P>
-        Users have the option to share their analytical outputs with certain
-        third party tools. You control whether to use these integrations or not.
-      </P>
-      <P>
-        In certain circumstances, we also share your Upload Data and analytical
-        results with other governmental, public health entities in accordance
-        with your organization’s policies and with applicable law. For example,
-        certain Users in California currently allow the California Department of
-        Public Health (&quot;CDPH&quot;) to access Upload Data and analytical
-        results from their Group.
-      </P>
-      <P>
-        In the unlikely event that we can no longer keep operating CZ GEN EPI or
-        believe that its purpose is better served by having another entity
-        operating it, we may transfer CZ GEN EPI and all data existing therein
-        (Upload Data, analytical outputs, and User Data) so that Users can
-        continue to be served. We will always let you know <B>before</B> this
-        happens, and you will have the option to delete your account and any
-        data you’ve uploaded. Should this occur, the entity to which we transfer
-        your data will be obliged to use it in a manner that is consistent with
-        this Privacy Policy and our{" "}
-        <NewTabLink href={ROUTES.TERMS}>Terms</NewTabLink>.
-      </P>
-      <P>
-        We may disclose Upload Data, analytical outputs, and/or User Data if we
-        believe in good faith that such disclosure is necessary (a) to comply
-        with our legal obligations or to respond to subpoenas or warrants served
-        on us; (b) to protect or defend our rights or property or those of
-        Users; and/or (c) to investigate or assist in preventing any violation
-        or potential violation of this Privacy Policy, or our{" "}
-        <NewTabLink href={ROUTES.TERMS}>Terms</NewTabLink>.
-      </P>
-    </>
-  );
-
-  const renderInformationProtectionPolicy = () => (
-    <>
-      <H2>
-        <Number>5.</Number>How We Protect the Information
-      </H2>
-      <P>
-        We use industry standard security measures to ensure the
-        confidentiality, integrity and availability of data uploaded into CZ GEN
-        EPI. This includes practices like encrypting connections to CZ GEN EPI
-        using TLS (encrypting data while in transit), hosting CZ GEN EPI on
-        leading cloud providers with robust physical security, and ensuring that
-        access to any personal data within CZ GEN EPI by CZIF, CZB, and CZI LLC
-        staff is limited to those staff who need access to operate the Service.
-      </P>
-      <P>
-        Security takes ongoing work and we will continue to monitor and adjust
-        our security measures as CZ GEN EPI develops. Please notify us
-        immediately at{" "}
-        <NewTabLink href="mailto:security@czgenepi.org">
-          security@czgenepi.org
-        </NewTabLink>{" "}
-        if you suspect your account has been compromised or are aware of any
-        other security issues relating to CZ GEN EPI.
+        We rely on consent as a legal basis when processing data through the use
+        of cookies when required under applicable law.
       </P>
     </>
   );
@@ -755,7 +727,7 @@ const PrivacyPolicy = (): JSX.Element => {
   const renderDataRetentionAndDeletionPolicy = () => (
     <>
       <H2>
-        <Number>6.</Number>How Long We Retain Data and Data Deletion
+        <Number>4.</Number>How Long We Retain Data and Data Deletion
       </H2>
       <P>
         We retain your personal data only as long as is reasonably necessary:
@@ -800,12 +772,101 @@ const PrivacyPolicy = (): JSX.Element => {
     </>
   );
 
+  const renderVendorAndThirdPartyPolicy = () => (
+    <>
+      <H2>
+        <Number>5.</Number>Vendors and Service Providers
+      </H2>
+      <P>
+        CZB, CZIF, and CZI LLC collaborate closely in order to build, design,
+        and operate CZ GEN EPI so that it can be as useful as possible to
+        researchers and the public health community. CZB and CZIF provide
+        scientific and data analysis leadership and CZI LLC focuses on
+        maintaining CZ GEN EPI’s infrastructure, security, and compliance. The
+        three parties are all data controllers for CZ GEN EPI and will all only
+        use data as described in this Privacy Policy.
+      </P>
+      <P>
+        We also use service providers, such as a database provider, analytics
+        providers, customer support providers, and survey providers to support
+        and improve the operation of CZ GEN EPI. These service providers are
+        data processors and their use is limited to the purposes disclosed in
+        this Privacy Policy. For more information about our use of service
+        providers, please see our{" "}
+        <NewTabLink href={ROUTES.PRIVACY_DATA_SHARING_FAQ}>
+          Privacy &amp; Data Sharing FAQ
+        </NewTabLink>
+        .
+      </P>
+      <P>
+        Users have the option to share their analytical outputs with certain
+        third party tools. You control whether to use these integrations or not.
+      </P>
+      <P>
+        In certain circumstances, we also share your Upload Data and analytical
+        results with other governmental, public health entities in accordance
+        with your organization’s policies and with applicable law. For example,
+        certain Users in California currently allow the California Department of
+        Public Health (&quot;CDPH&quot;) to access Upload Data and analytical
+        results from their Group.
+      </P>
+      <P>
+        In the unlikely event that we can no longer keep operating CZ GEN EPI or
+        believe that its purpose is better served by having another entity
+        operating it, we may transfer CZ GEN EPI and all data existing therein
+        (Upload Data, analytical outputs, and User Data) so that Users can
+        continue to be served. We will always let you know <B>before</B> this
+        happens, and you will have the option to delete your account and any
+        data you’ve uploaded prior to any such transfer. Should this occur, the
+        entity to which we transfer your data will be obliged to use it in a
+        manner that is consistent with this Privacy Policy and our{" "}
+        <NewTabLink href={ROUTES.TERMS}>Terms</NewTabLink>.
+      </P>
+      <P>
+        We may disclose Upload Data, analytical outputs, and/or User Data if we
+        believe in good faith that such disclosure is necessary (a) to comply
+        with our legal obligations or to respond to subpoenas or warrants served
+        on us; (b) to protect or defend our rights or property or those of
+        Users; and/or (c) to investigate or assist in preventing any violation
+        or potential violation of this Privacy Policy, or our{" "}
+        <NewTabLink href={ROUTES.TERMS}>Terms</NewTabLink>.
+      </P>
+    </>
+  );
+
+  const renderInformationProtectionPolicy = () => (
+    <>
+      <H2>
+        <Number>6.</Number>How We Protect the Information
+      </H2>
+      <P>
+        We use industry standard security measures to ensure the
+        confidentiality, integrity and availability of data uploaded into CZ GEN
+        EPI. This includes practices like encrypting connections to CZ GEN EPI
+        using TLS (encrypting data while in transit), hosting CZ GEN EPI on
+        leading cloud providers with robust physical security, and ensuring that
+        access to any personal data within CZ GEN EPI by CZIF, CZB, and CZI LLC
+        staff is limited to those staff who need access to operate the Service.
+      </P>
+      <P>
+        Security takes ongoing work and we will continue to monitor and adjust
+        our security measures as CZ GEN EPI develops. Please notify us
+        immediately at{" "}
+        <NewTabLink href="mailto:security@czgenepi.org">
+          security@czgenepi.org
+        </NewTabLink>{" "}
+        if you suspect your account has been compromised or are aware of any
+        other security issues relating to CZ GEN EPI.
+      </P>
+    </>
+  );
+
   const renderUserDataChoicesPolicy = () => (
     <>
       <H2>
         <Number>7.</Number>Choices About Your Data
       </H2>
-      <P>Users have the following choices:</P>
+      <P>Users have the following choices and rights:</P>
       <List>
         <ListItem>
           Users are able to request the deletion of User Data that constitutes
@@ -825,6 +886,11 @@ const PrivacyPolicy = (): JSX.Element => {
           CZ GEN EPI.
         </ListItem>
         <ListItem>
+          Users have the right to receive notice of our personal information
+          collection, use, retention, and disclosure practices at or before
+          collection of personal information.
+        </ListItem>
+        <ListItem>
           <span>
             If you have any questions about our processing of any data, please
             contact us at{" "}
@@ -835,13 +901,75 @@ const PrivacyPolicy = (): JSX.Element => {
           </span>
         </ListItem>
       </List>
+      <P>
+        We may ask for additional information that allows us to reasonably
+        verify that you are the person about whom we collected personal
+        information. If you would like an authorized agent to make a request for
+        you, have that agent email{" "}
+        <NewTabLink href="mailto:privacy@czgenepi.org">
+          privacy@czgenepi.org
+        </NewTabLink>{" "}
+        with information that is sufficient for us to verify that the authorized
+        agent is acting on your behalf.
+      </P>
+      <P>
+        If you would like to appeal a decision with respect to a request to
+        exercise any of these rights, please email us at{" "}
+        <NewTabLink href="mailto:privacy@czgenepi.org">
+          privacy@czgenepi.org
+        </NewTabLink>{" "}
+        and explain the basis for your appeal.
+      </P>
+      <P>
+        If your data is subject to the EU or UK data protection laws (e.g.,
+        GDPR) and you wish to raise a concern about our use of your information
+        (and without prejudice to any other rights you may have), you have the
+        right to do so with your local supervisory authority or by emailing us
+        at{" "}
+        <NewTabLink href="mailto:privacy@czgenepi.org">
+          privacy@czgenepi.org
+        </NewTabLink>
+        .
+      </P>
     </>
   );
 
-  const renderDataLocation = () => (
+  const renderChildren = () => (
     <>
       <H2>
-        <Number>8.</Number>Data Location
+        <Number>8.</Number>Children
+      </H2>
+      <P>
+        Please note: CZ GEN EPI is not designed for or directed toward children
+        under the age of 16 and we do not have actual knowledge that we have
+        sold or shared the personal information of users under 16 years of age.
+        If we become aware that we have the information of such children
+        collected through CZ GEN EPI, we will promptly delete it.
+      </P>
+    </>
+  );
+
+  const renderDoNotTrack = () => (
+    <>
+      <H2>
+        <Number>9.</Number>Do Not Track
+      </H2>
+      <P>
+        We don’t share personal data with third parties for their direct
+        marketing purposes or behavioral advertising, nor do we support any Do
+        Not Track signals since there’s currently no standard for how online
+        services respond to those signals. As standards develop, we may
+        establish policies for responding to DNT signals that we would describe
+        in this Privacy Notice. You can learn more about DNT{" "}
+        <NewTabLink href="https://allaboutdnt.com/">here</NewTabLink>.
+      </P>
+    </>
+  );
+
+  const renderDataTransfers = () => (
+    <>
+      <H2>
+        <Number>10.</Number>Data Transfers
       </H2>
       <P>
         CZ GEN EPI is a US-based service. If you want to use CZ GEN EPI, you
@@ -851,13 +979,188 @@ const PrivacyPolicy = (): JSX.Element => {
         States, and use technical infrastructure in the United States to deliver
         the Services to you.
       </P>
+      <P>
+        In accordance with the contract between us and our Users, we need to
+        transfer personal data to the United States and to other jurisdictions
+        as necessary to provide the Services. Please note that the privacy
+        protections with respect to your information in these countries may not
+        be the same as in your home country.
+      </P>
+    </>
+  );
+
+  const renderCaliforniaResidents = () => (
+    <>
+      <H2>
+        <Number>11.</Number>Additional Information for California Residents
+      </H2>
+      <P>
+        The California Consumer Privacy Act (“CCPA”) requires certain businesses
+        to give California residents a number of rights regarding their personal
+        information. We are offering these rights to you, including the right to
+        have your personal information deleted (subject to certain exceptions),
+        the right to change or correct your personal information, the right to
+        limit the use or disclosure of your sensitive personal information (if
+        applicable), the right to access your personal information, the right to
+        opt-out of the “selling” or “sharing” of personal information (if
+        applicable), and the right not to be discriminated against for
+        exercising these rights.
+      </P>
+      <P>
+        These rights, and how to exercise them, are described in more detail in
+        the Section titled “Choices about Your Data” of this Policy Policy. In
+        addition to these rights, we give you a right to request the following
+        information about your personal information that we have collected in
+        the past 12 months:
+      </P>
+      <P>
+        <B>The Right to Know.</B> This right allows you to request the following
+        information about the personal information that we’ve collected about
+        you in the past 12 months:
+      </P>
+      <List>
+        <ListItem>
+          <span>
+            <B>Information about Data Collection</B>
+            <List>
+              <ListItem>
+                The categories of personal information that have been collected
+                about you.
+              </ListItem>
+              <ListItem>
+                The categories of sources from which we have collected personal
+                information.
+              </ListItem>
+              <ListItem>
+                The business purpose for which we have collected personal
+                information.
+              </ListItem>
+            </List>
+          </span>
+        </ListItem>
+        <ListItem>
+          <span>
+            <B>Information about Data Disclosure</B>
+            <List>
+              <ListItem>
+                The categories of personal information, if any, that have been
+                sold (if applicable), shared, or disclosed for a business
+                purpose to third parties. Note: CZ GEN EPI does not sell your
+                personal information nor will we share your personal information
+                for behavioral advertising purposes.
+              </ListItem>
+              <ListItem>
+                The categories of third parties to whom personal information was
+                sold, shared, or disclosed for a business purpose.
+              </ListItem>
+              <ListItem>
+                Identification of the specific business purpose for disclosing
+                the consumer’s personal information.
+              </ListItem>
+            </List>
+          </span>
+        </ListItem>
+      </List>
+      <P>
+        We have described in fuller detail in this Privacy Policy the personal
+        information that we collect, how we use, and disclose it, but provide
+        the following additional disclosure:
+      </P>
+      <P>
+        <B>Information about Data Collection</B>
+      </P>
+      <List>
+        <ListItem>
+          <span>
+            <B>Information we collect.</B> We have collected the following
+            categories of personal information from consumers within the past 12
+            months: (1) identifiers; (2) professional or employment-related
+            information; (3) internet or other electronic network activity
+            within CZ GEN EPI; (4) geolocation data; (5) inferences drawn from
+            your internet or other electronic network activity within CZ GEN
+            EPI; and (6) information provided within survey responses.
+          </span>
+        </ListItem>
+        <ListItem>
+          <span>
+            <B>Sources of information.</B> We obtain these categories of
+            personal information directly from you and from your browser or
+            device as you use the Services.
+          </span>
+        </ListItem>
+        <ListItem>
+          <span>
+            <B>Purposes of collection.</B> We collect personal information for
+            one or more of the following business purposes as described in
+            Sections 1 through 3 above.
+          </span>
+        </ListItem>
+      </List>
+      <P>
+        <B>Information about Data Disclosure</B>
+      </P>
+      <List>
+        <ListItem>
+          <span>
+            <B>Information we disclose.</B> We have disclosed the following
+            categories of personal information within the past 12 months: (1)
+            identifiers; (2) professional or employment-related information; (3)
+            internet or other electronic network activity within CZ GEN EPI; (4)
+            geolocation data; (5) inferences drawn from your internet or other
+            electronic network activity within CZ GEN EPI; and (6) information
+            provided within survey responses.
+          </span>
+        </ListItem>
+        <ListItem>
+          <span>
+            <B>Third parties to whom we disclose.</B> The categories of third
+            parties to whom we have disclosed this personal information are
+            described in the Section titled “Vendors and Service Providers” in
+            this Privacy Policy.
+          </span>
+        </ListItem>
+        <ListItem>
+          <span>
+            <B>Purposes of disclosure.</B> We disclose the personal information
+            we collect about you for one or more of the following business
+            purposes as described in Sections 1 through 3 above.
+          </span>
+        </ListItem>
+      </List>
+    </>
+  );
+
+  const renderAdditionalResidents = () => (
+    <>
+      <H2>
+        <Number>12.</Number>Additional Information for Residents of Virginia,
+        Colorado, Connecticut, and Utah
+      </H2>
+      <P>
+        Virginia, Colorado, Connecticut, and Utah also have adopted privacy laws
+        that give consumers certain rights, including the right to confirm
+        whether controllers are processing the consumer’s personal data, the
+        right to access that data, the right to obtain a copy of that data, the
+        right to correct inaccuracies in that data, and the right to delete that
+        data. As discussed above in the Section titled “Choices About Your
+        Data,” we provide these rights to all consumers, regardless of where
+        they reside.
+      </P>
+      <P>
+        Additionally, these four states have adopted rights to opt-out of: (1)
+        targeted advertising; (2) the sale of personal data; and (3) profiling
+        in furtherance of decisions that produce legal or similarly significant
+        effects concerning the consumer. We do not sell your data, use it for
+        targeted advertising or to profile you in furtherance of decisions that
+        produce legal or similarly significant effects.
+      </P>
     </>
   );
 
   const renderContactInfo = () => (
     <>
       <H2>
-        <Number>9.</Number>How to Contact Us
+        <Number>13.</Number>How to Contact Us
       </H2>
       <P>
         If you have any questions, comments, or concerns with this Privacy
@@ -873,11 +1176,12 @@ const PrivacyPolicy = (): JSX.Element => {
   const renderChangesToPrivacyNotice = () => (
     <>
       <H2>
-        <Number>10.</Number>Changes to This Privacy Notice
+        <Number>14.</Number>Changes to This Privacy Notice
       </H2>
       <P>
+        This Privacy Policy was last updated on the “Last Updated” date above.
         We may update this Privacy Policy from time to time and will provide you
-        with notice of material updates before they become effective.
+        with notice of any material updates before they become effective.
       </P>
     </>
   );
@@ -889,14 +1193,19 @@ const PrivacyPolicy = (): JSX.Element => {
         <NarrowContainer>
           {renderIntro()}
           {renderSummaryTable()}
+          {renderDataCZGenEpiCollects()}
           {renderUploadDataPolicy()}
           {renderReportDataPolicy()}
           {renderVisitorAndUserDataPolicy()}
+          {renderDataRetentionAndDeletionPolicy()}
           {renderVendorAndThirdPartyPolicy()}
           {renderInformationProtectionPolicy()}
-          {renderDataRetentionAndDeletionPolicy()}
           {renderUserDataChoicesPolicy()}
-          {renderDataLocation()}
+          {renderChildren()}
+          {renderDoNotTrack()}
+          {renderDataTransfers()}
+          {renderCaliforniaResidents()}
+          {renderAdditionalResidents()}
           {renderContactInfo()}
           {renderChangesToPrivacyNotice()}
         </NarrowContainer>
